refactor: move MapperRegistry to mapper.support package
