+++ nextstep-jdbc/src/main/java/nextstep/jdbc/mapper/support/MapperRegistry.java
@@ -1,5 +1,8 @@
-package nextstep.jdbc.mapper;
+package nextstep.jdbc.mapper.support;
 
+import nextstep.jdbc.mapper.ListMapper;
+import nextstep.jdbc.mapper.ObjectMapper;
+import nextstep.jdbc.mapper.UserMapper;
 import slipp.domain.User;
 
 import java.util.HashMap;
