+++ slipp/src/main/java/slipp/dao/UserDao.java
@@ -1,6 +1,6 @@
 package slipp.dao;
 
-import nextstep.jdbc.mapper.MapperRegistry;
+import nextstep.jdbc.mapper.support.MapperRegistry;
 import nextstep.jdbc.template.JdbcTemplate;
 import slipp.domain.User;
 import slipp.support.db.ConnectionManager;
